+++ Sources/KeyboardKit/Callouts/Callouts.swift
@@ -11,7 +11,7 @@ import Foundation
 /**
  This namespace contains callout-related types.
  
- This namespace doesn't contain services or observable types,
- or types that need to be exposed to the library's top level.
+ The namespace doesn't contain protocol, observable types or
+ open classes, not types that are meant to be top-level ones.
  */
 public struct Callouts {}