+++ Sources/KeyboardKit/Proxy/Proxy.swift
@@ -1,17 +1,20 @@
 //
-//  Emojis.swift
+//  Proxy.swift
 //  KeyboardKit
 //
-//  Created by Daniel Saidi on 2023-09-20.
+//  Created by Daniel Saidi on 2023-09-19.
 //  Copyright © 2023 Daniel Saidi. All rights reserved.
 //
 
 import Foundation
 
 /**
- This namespace contains emojis-related types.
+ This namespace contains proxy-related types.
  
  The namespace doesn't contain protocol, observable types or
  open classes, not types that are meant to be top-level ones.
+ 
+ For now, this namespace will only have things in it when it
+ is part of the KeyboardKit Pro build.
  */
-public struct Emojis {}
+public struct Proxy {}