+++ Sources/KeyboardKit/Proxy/UITextDocumentProxy+Delete.swift
@@ -11,17 +11,13 @@ import UIKit
 
 public extension UITextDocumentProxy {
 
-    /**
-     Delete backwards a certain range.
-     */
+    /// Delete backwards a certain range.
     func deleteBackward(range: Keyboard.BackspaceRange) {
         guard let text = deleteBackwardText(for: range) else { return deleteBackward() }
         deleteBackward(times: text.count)
     }
     
-    /**
-     Delete backwards a certain number of times.
-     */
+    /// Delete backwards a certain number of times.
     func deleteBackward(times: Int) {
         for _ in 0..<times { deleteBackward() }
     }